refactor llm profile
